finished UI fixes
