+++ App/__tests__/RemoveGroupMaster.test.tsx
@@ -1,6 +1,6 @@
 import { render } from "@testing-library/react";
 import { BrowserRouter } from "react-router-dom";
-import RemovingMasterFromGroup from "../src/pages/RemoveGroupMaster";
+import RemovingMasterFromGroup from "../src/pages/groups/RemoveGroupMaster";
 
 describe("Add User to Group Test Case", () => {
     it("validate function should render all elements ", () => {
